requirements/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
bleach==6.0.0
blessed==1.20.0
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
decorator==5.1.1
Django==4.2.2
django-environ==0.10.0
django-extensions==3.2.3
django-ninja==0.22.2
django-picklefield==3.1
django-q==1.3.9
executing==1.2.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.25.0
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.1.2
priority==2.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.10
Pygments==2.15.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
redis==3.5.3
service-identity==23.1.0
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
toml==0.10.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.0
tzdata==2023.3
ujson==5.8.0
wcwidth==0.2.6
webencodings==0.5.1
wheel==0.40.0
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
